requirements.txt
tensorflow==2.19.0
keras==3.9.2
numpy==2.1.3
matplotlib==3.10.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
pandas==2.2.2
h5py==3.11.0
Pillow==10.3.0
opencv-python==4.9.0.80
nibabel==5.2.1
SimpleITK==2.3.1
tqdm==4.66.2